+++ internal/cli/checkout.go
@@ -21,7 +21,7 @@ func newCheckoutCmd() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			mods, err := module.Resolve(cfg, flagPath, flagProject, flagNoRoot)
+			mods, err := module.Resolve(cfg, flagPath, flagProject, flagNoRoot, flagExcept)
 			if err != nil {
 				return err
 			}