+++ internal/cli/pull.go
@@ -27,7 +27,7 @@ func newPullCmd() *cobra.Command {
 			if err != nil {
 				return err
 			}
-			mods, err := module.Resolve(cfg, flagPath, flagProject, flagNoRoot)
+			mods, err := module.Resolve(cfg, flagPath, flagProject, flagNoRoot, flagExcept)
 			if err != nil {
 				return err
 			}